backend/requirements.txt
requests 
beautifulsoup4 
newspaper3k 
tqdm
chromadb 
sentence-transformers
fastapi 
uvicorn 
instructor 
pydantic 
unstructured 
duckdb
onnxruntime
google-cloud-aiplatform
langchain-community
pydantic<2.0
